test: relocate the attention cases to agent/attention-registry.test.ts
Five cases move to `src/backend/agent/attention-registry.test.ts`, named for
the boundary they now cover: the Codex approval round trip with the runtime
attention budget (the whole of `agent-service.approvals.test.ts` after PR 2
took the six hosted-site cases), and four from
`agent-service.questions.test.ts` — the Computer Use elicitation, the
default-mode `ask_user` tool, the paused browser tool call, and the legacy
approval that must still be interactive and must be gone after `stop()`.

Relocation only: every case still constructs a real `AgentService` through
`agent-service-test-harness.ts` and asserts through public events, the
runtime snapshot and the provider responses. Not one assertion changed —
the two files' post-move diffs touch nothing but the import blocks, which
Biome rewrote for the names each file no longer uses.

Four cases stay in `agent-service.questions.test.ts` because their
consequence crosses a domain the registry does not own: the question prompt
summary must not become the completed assistant reply (conversation
projection), prompts from a healthy provider survive another provider's exit
(provider lifecycle), and the two automatic-memory cases belong to the
controller PR 6 extracts.

`agent-service.approvals.test.ts` is deleted rather than left with an empty
`describe` — its last case moved, none was dropped. Nine cases before, nine
after; the four remaining agent-service shards and all five
`src/backend/agent/*.test.ts` files pass (89 tests).
